src/Exercise3/Exercise3/requirements.txt
pyzmq==16.0.*
multiprocessing_logging>=0.2.4
